1.实现果蔬编辑、删除 2.实现果蔬类别显示、编辑 3.解决日志只显示一页的bug 4.发现当果蔬类别被修改时，先有的果蔬信息里的类别还是以前的名字的bug

+++ src/main/java/com/zl/mapper/LogMapper.java
@@ -46,7 +46,7 @@ public interface LogMapper {
     * @Author: ZhuLin
     * @Date: 2019/1/18 
     */ 
-    Long getLogCount();
+    Integer getLogCount();
     
     /*** 
     * @Description: 清除超过五天日志记录
